Moved `FoundationDataBuilder.post` to `../network`.
